requirements.txt
about-time==4.2.1
absl-py==2.3.1
accelerate==1.10.1
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
alive-progress==3.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
argcomplete==3.6.2
attrs==25.1.0
audioop-lts==0.2.2
audioread==3.0.1
autograd==1.8.0
beautifulsoup4==4.13.5
blivet==3.12.1
blivet-gui==2.6.0
boto3==1.40.26
botocore==1.40.26
Brlapi==0.8.6
Brotli==1.1.0
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
cma==4.3.0
coloredlogs==15.0.1
contourpy==1.3.3
crypt_r==3.13.1
cssselect==1.2.0
cupshelpers==1.0
cycler==0.12.1
dasbus==1.7
datasets==4.1.0
dbus-python==1.3.2
dbus_next==0.2.3
decorator==5.2.1
Deprecated==1.2.18
dill==0.4.0
distro==1.9.0
dnf==4.23.0
fedora-third-party==0.10
file-magic==0.4.0
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
GitPython==3.1.45
graphemeu==0.7.2
grpcio==1.74.0
hf-xet==1.1.10
huggingface-hub==0.34.5
humanfriendly==10.0
humanize==4.12.0
hydra-core==1.3.2
i3ipc==2.2.1
idna==3.10
initial-setup==0.3.101
iso639==0.1.4
Jinja2==3.1.6
jiwer==4.0.0
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.9
langtable==0.0.69
lazy_loader==0.4
libcomps==0.1.22
libdnf==0.74.0
librosa==0.11.0
llvmlite==0.44.0
louis==3.33.0
lxml==5.3.2
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
mercurial==6.9.5
ml_dtypes==0.5.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.6.4
multiprocess==0.70.16
mutagen==1.47.0
natsort==8.4.0
netifaces==0.11.0
networkx==3.4.2
nftables==0.1
ninja==1.13.0
nncf==2.18.0
numba==0.61.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
nwg-displays==0.3.26
nwg-panel==0.10.12
olefile==0.47
omegaconf==2.3.0
onnx==1.19.0
onnxruntime==1.22.1
openvino==2025.3.0
openvino-telemetry==2025.2.0
optimum==1.27.0
optimum-quanto==0.2.7
packaging==24.2
pandas==2.3.2
perf==0.1
pexpect==4.9.0
pid==2.2.3
pillow==11.1.0
platformdirs==4.4.0
ply==3.11
pooch==1.8.2
productmd==1.45
progressbar2==4.5.0
propcache==0.3.2
protobuf==6.32.1
psutil==6.1.1
ptyprocess==0.7.0
pwquality==1.4.5
pyarrow==21.0.0
PyAudio==0.2.13
pycairo==1.25.1
pycparser==2.20
pycryptodomex==3.23.0
pycups==2.0.4
pydantic==2.11.9
pydantic_core==2.33.2
pydot==3.0.4
pyenchant==3.2.2
pygdbmi==0.11.0.0
Pygments==2.19.2
PyGObject==3.50.0
pyinotify==0.9.6
pykickstart==3.62
pymoo==0.6.1.5
pyparsing==3.2.4
pyparted==3.13.0
pyquery==2.0.1
PySocks==1.7.1
python-augeas==1.2.0
python-dateutil==2.8.2
python-linux-procfs==0.7.3
python-meh==0.52
python-utils==3.9.0
python-xlib==0.33
pytz==2025.2
pyudev==0.24.3
pyxdg==0.27
PyYAML==6.0.2
pyynl @ file:///builddir/build/BUILD/kernel-6.16.7-build/kernel-6.16.7/linux-6.16.7-200.fc42.x86_64/tools/net/ynl
RapidFuzz==3.11.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
requests-file==2.0.0
requests-ftp==0.3.1
rich==14.1.0
rpds-py==0.25.0
rpm==4.20.1
rpmautospec==0.8.1
rpmautospec-core==0.1.5
s3transfer==0.13.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
selinux @ file:///builddir/build/BUILD/libselinux-3.8-build/libselinux-3.8/src
sentry-sdk==2.21.0
sepolicy @ file:///builddir/build/BUILD/policycoreutils-3.8-build/selinux-3.8/python/sepolicy
setools==4.5.1
setroubleshoot @ file:///builddir/build/BUILD/setroubleshoot-3.3.35-build/setroubleshoot-3.3.35/src
setuptools==74.1.3
simpleaudio==1.0.4
simpleline==1.9.0
six==1.17.0
smmap==5.0.2
sos==4.10.0
soundfile==0.13.1
soupsieve==2.7
soxr==1.0.0
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
sympy==1.14.0
systemd-python==235
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
tokenizers==0.21.4
torch==2.8.0
torchaudio==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.53.3
triton==3.4.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.3.0
wandb==0.21.4
websockets==14.1
Werkzeug==3.1.3
wrapt==1.17.3
xkbregistry==0.3
xxhash==3.5.0
yarl==1.20.1
yt-dlp==2025.9.5
